Conflicts:
	crm-parent/crm-manager/src/main/java/com/zijincaifu/crm/manage/controller/channel/ChannelController.java
	crm-parent/crm-manager/src/main/resources/spring/applicationContext-shiro.xml
	crm-parent/crm-manager/src/main/webapp/WEB-INF/views/fm/manage/base.html
	crm-parent/crm-manager/src/main/webapp/WEB-INF/views/fm/manage/channel/channelList.html
	crm-parent/crm-manager/src/main/webapp/WEB-INF/views/fm/manage/customer/addCustomer.html
	crm-parent/crm-manager/src/main/webapp/WEB-INF/views/fm/manage/customer/invest.html
	crm-parent/crm-manager/src/main/webapp/WEB-INF/views/fm/manage/index.html
	crm-parent/crm-manager/src/main/webapp/WEB-INF/views/fm/manage/login.html
	crm-parent/crm-manager/src/main/webapp/WEB-INF/views/fm/manage/personnel/personnelAdd.html
	crm-parent/crm-manager/src/main/webapp/WEB-INF/views/fm/manage/personnel/personnelList.html
	crm-parent/crm-service/src/main/java/com/zijincaifu/service/customer/ICustomerService.java
	crm-parent/crm-service/src/main/java/com/zijincaifu/service/customer/IInvestItemService.java
	crm-parent/crm-service/src/main/java/com/zijincaifu/service/impl/customer/CustomerServiceImpl.java
	crm-parent/crm-service/src/main/java/com/zijincaifu/service/impl/customer/InvestItemServiceImpl.java

+++ crm-parent/crm-manager/src/main/java/com/zijincaifu/crm/manage/controller/customer/CustomerController.java
@@ -5,6 +5,7 @@ import java.io.PrintWriter;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
@@ -12,6 +13,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import org.apache.shiro.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
@@ -20,17 +22,23 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import com.sxj.cache.manager.CacheLevel;
+import com.sxj.cache.manager.HierarchicalCacheManager;
 import com.sxj.util.common.StringUtils;
 import com.sxj.util.exception.WebException;
 import com.sxj.util.logger.SxjLogger;
 import com.zijincaifu.crm.entity.customer.CustomerEntity;
+import com.zijincaifu.crm.entity.customer.TrackRecordEntity;
 import com.zijincaifu.crm.entity.personnel.PersonnelEntity;
 import com.zijincaifu.crm.entity.system.AreaEntity;
 import com.zijincaifu.crm.enu.customer.CustomerLevelEnum;
 import com.zijincaifu.crm.manage.controller.BaseController;
+import com.zijincaifu.crm.model.customer.InvestItemModel;
 import com.zijincaifu.model.customer.CustomerQuery;
 import com.zijincaifu.model.personnel.PersonnelQuery;
 import com.zijincaifu.service.customer.ICustomerService;
+import com.zijincaifu.service.customer.IInvestItemService;
+import com.zijincaifu.service.customer.ITrackRecordService;
 import com.zijincaifu.service.personnel.IPersonnelService;
 import com.zijincaifu.service.system.IAreaService;
 
@@ -48,17 +56,57 @@ public class CustomerController extends BaseController
     @Autowired
     private IPersonnelService personnelService;
     
+    @Autowired
+    private IInvestItemService investItemService;
+    
+    @Autowired
+    private ITrackRecordService trackRecordService;
+    
     @RequestMapping("/query")
     public String query(CustomerQuery query, ModelMap map) throws WebException
     {
         try
         {
             query.setPagable(true);
+            query.setShowCount(15);
+            boolean isAdmin = true;
+            if (SecurityUtils.getSubject() != null
+                    && !SecurityUtils.getSubject().hasRole("4"))
+            {
+                PersonnelEntity user = getLoginInfo();
+                query.setEmployeId(user.getUid());
+                isAdmin = false;
+            }
             List<AreaEntity> provinceList = areaService.getChildrenAreas("0");
             List<AreaEntity> cityList = areaService.getChildrenAreas("32");
             map.put("provinceList", provinceList);
             map.put("cityList", cityList);
             List<CustomerEntity> list = customerService.queryCustomer(query);
+            for (Iterator<CustomerEntity> iterator = list.iterator(); iterator.hasNext();)
+            {
+                CustomerEntity customerEntity = iterator.next();
+                if (isAdmin)
+                {
+                    Object isNew = HierarchicalCacheManager.get(CacheLevel.REDIS,
+                            "CRM_CUSTOMER_LEVEL_AD",
+                            customerEntity.getCustomerId());
+                    if (isNew != null)
+                    {
+                        customerEntity.setIsNew((String) isNew);
+                    }
+                }
+                else
+                {
+                    Object isNew = HierarchicalCacheManager.get(CacheLevel.REDIS,
+                            "CRM_CUSTOMER_LEVEL",
+                            customerEntity.getCustomerId());
+                    if (isNew != null)
+                    {
+                        customerEntity.setIsNew((String) isNew);
+                    }
+                }
+                
+            }
             CustomerLevelEnum[] levels = CustomerLevelEnum.values();
             map.put("list", list);
             map.put("levels", levels);
@@ -70,6 +118,40 @@ public class CustomerController extends BaseController
             SxjLogger.error("查询客户信息错误", e, this.getClass());
             throw new WebException("查询客户信息错误", e);
         }
+    }
+    
+    @RequestMapping("/changeNew")
+    public @ResponseBody Map<String, Object> changeNew(String customerId)
+            throws WebException
+    {
+        Map<String, Object> map = new HashMap<String, Object>();
+        try
+        {
+            if (SecurityUtils.getSubject() != null
+                    && !SecurityUtils.getSubject().hasRole("4"))
+            {
+                HierarchicalCacheManager.set(CacheLevel.REDIS,
+                        "CRM_CUSTOMER_LEVEL",
+                        customerId,
+                        "0");
+                map.put("isNew", "0");
+            }
+            else
+            {
+                HierarchicalCacheManager.set(CacheLevel.REDIS,
+                        "CRM_CUSTOMER_LEVEL_AD",
+                        customerId,
+                        "0");
+                map.put("isNew", "0");
+            }
+            
+        }
+        catch (Exception e)
+        {
+            SxjLogger.error("查询客户信息错误", e, this.getClass());
+            map.put("isNew", "1");
+        }
+        return map;
     }
     
     @RequestMapping("/toAdd")
@@ -129,11 +211,15 @@ public class CustomerController extends BaseController
                 throw new WebException("产品不能为空");
             }
             //TODO 设置员工
-            //customer.setEmployeId(login.getUid());
+            customer.setEmployeId(login.getUid());
             customer.setChannelId("AAA001");
             customer.setCreateTime(new Date());
             customer.setLevel(CustomerLevelEnum.NEW);
             boolean isAdd = customerService.addCustomer(customer, productId);
+            HierarchicalCacheManager.set(CacheLevel.REDIS,
+                    "CRM_CUSTOMER_LEVEL_AD",
+                    customer.getCustomerId(),
+                    "1");
             map.put("isOK", true);
             map.put("isAdd", isAdd);
         }
@@ -162,8 +248,9 @@ public class CustomerController extends BaseController
             {
                 throw new WebException("客户信息不能为空");
             }
-            customerService.modifyCustomer(customer);
+            customerService.modifyCustomer(customer,getLoginInfo().getUid());
             map.put("isOK", true);
+            map.put("customer", customer);
         }
         catch (Exception e)
         {
@@ -184,7 +271,7 @@ public class CustomerController extends BaseController
         {
             CustomerEntity customer = customerService.getCustomer(customerId);
             customer.setLevel(level);
-            customerService.updateCustomer(customer);
+            customerService.updateCustomer(customer,getLoginInfo().getUid());
             map.put("isOK", true);
         }
         catch (Exception e)
@@ -254,9 +341,18 @@ public class CustomerController extends BaseController
             else
             {
                 CustomerEntity customer = customerService.getCustomer(customerId);
+                if(customer.getEmployeIdHistory()==null){
+                    customer.setEmployeIdHistory(customer.getEmployeId());
+                }else{
+                    customer.setEmployeIdHistory(customer.getEmployeIdHistory()+","+customer.getEmployeId());
+                }
                 customer.setEmployeId(personnelId);
-                customerService.updateCustomer(customer);
+                customerService.updateCustomer(customer,"");
                 map.put("result", "1");
+                HierarchicalCacheManager.set(CacheLevel.REDIS,
+                        "CRM_CUSTOMER_LEVEL",
+                        customer.getCustomerId(),
+                        "1");
             }
         }
         catch (Exception e)
@@ -275,9 +371,29 @@ public class CustomerController extends BaseController
         Map<String, Object> map = new HashMap<String, Object>();
         try
         {
-//            CustomerEntity customer = customerService.getCustomer(customerId);
-            customerService.deleteCustomer(customerId);
-            map.put("isOK", true);
+            CustomerEntity customer = customerService.getCustomer(customerId);
+            List<InvestItemModel> invests = investItemService.queryItems(customerId);
+            List<TrackRecordEntity> tracks = trackRecordService.query(customerId);
+            if (customer.getLevel().getId() != 0)
+            {
+                map.put("isOK", false);
+                map.put("error", "该用户不是新客户,不可删除");
+            }
+            else if (invests.size() > 1)
+            {
+                map.put("isOK", false);
+                map.put("error", "该用户下存在不止一条投资记录,不可删除");
+            }
+            else if (tracks.size() != 0)
+            {
+                map.put("isOK", false);
+                map.put("error", "该用户下存在跟踪记录,不可删除");
+            }
+            else
+            {
+                customerService.deleteCustomer(customerId);
+                map.put("isOK", true);
+            }
         }
         catch (Exception e)
         {
@@ -287,4 +403,18 @@ public class CustomerController extends BaseController
         }
         return map;
     }
+    
+    @RequestMapping("info")
+    public String getCustomerInfo(String customerId, ModelMap map)
+            throws WebException
+    {
+        CustomerEntity customer = customerService.getCustomer(customerId);
+        List<AreaEntity> provinceList = areaService.getChildrenAreas("0");
+        List<AreaEntity> cityList = areaService.getChildrenAreas("32");
+        map.put("provinceList", provinceList);
+        map.put("cityList", cityList);
+        map.put("model", customer);
+        return "manage/customer/customerInfo";
+    }
+    
 }
